Add validate-golint target and script …
… and *lint* the latest piece of code.
