+++ configuration.go
@@ -1,12 +1,16 @@
 package main
 
 import (
+	fmtlog "log"
 	"time"
 
+	"github.com/BurntSushi/toml"
 	"github.com/emilevauge/traefik/provider"
 	"github.com/emilevauge/traefik/types"
 )
 
+// GlobalConfiguration holds global configuration (with providers, etc.).
+// It's populated from the traefik configuration file passed as an argument to the binary.
 type GlobalConfiguration struct {
 	Port                      string
 	GraceTimeOut              int64
@@ -25,6 +29,7 @@ type GlobalConfiguration struct {
 	Boltdb                    *provider.BoltDb
 }
 
+// NewGlobalConfiguration returns a GlobalConfiguration with default values.
 func NewGlobalConfiguration() *GlobalConfiguration {
 	globalConfiguration := new(GlobalConfiguration)
 	// default values
@@ -34,6 +39,15 @@ func NewGlobalConfiguration() *GlobalConfiguration {
 	globalConfiguration.ProvidersThrottleDuration = time.Duration(2 * time.Second)
 
 	return globalConfiguration
+}
+
+// LoadFileConfig returns a GlobalConfiguration from reading the specified file (a toml file).
+func LoadFileConfig(file string) *GlobalConfiguration {
+	configuration := NewGlobalConfiguration()
+	if _, err := toml.DecodeFile(file, configuration); err != nil {
+		fmtlog.Fatalf("Error reading file: %s", err)
+	}
+	return configuration
 }
 
 type configs map[string]*types.Configuration
