+++ version.go
@@ -1,6 +1,8 @@
 package main
 
 var (
-	Version   = ""
+	// Version holds the current version of traefik.
+	Version = ""
+	// BuildDate holds the build date of traefik.
 	BuildDate = ""
 )
